fix : 사용자 입력 모달창 버그 개선
